[ty] Move attribute assignment validation into builder module
